Retrieve DNS from k8s only on orchestrator creation
